Add Docker containerization and improve environment loading
Adds a basic Dockerfile and docker-compose setup to make local development a bit easier and remove the need to e.g have ffmpeg installed.

Also makes a small tweak to the logging to make it so that we only even bother trying to set up Axiom things if we're in a non-local environment. This means developers would no longer require functional Axiom credentials in their env file to start the application.

Also logging-related, I stopped throwing an error if a .env file doesn't exist, as it wouldn't in production environments. I retained a small info message about what the system is doing, but moved the logging of it to be after the handler set-up so it is properly dealt with and forwarded to Axiom if that environment is configured